test: add CopilotPluginCommonTest class for common test setup
