+++ packages/tests/src/e2e/copilotplugin/CopilotPluginWithOAuth.tests.ts
@@ -1,30 +1,18 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
-import { Capability } from "../../utils/constants";
-import { CaseFactory } from "../caseFactory";
+
+/**
+ * @author Hui Miao <huimiao@microsoft.com>
+ */
+
 import { ProgrammingLanguage } from "@microsoft/teamsfx-core";
+import { CopilotPluginCommonTest } from "./CopilotPluginCommonTest";
 
-class CopilotPluginOAuthTestCase extends CaseFactory {}
-
-const copilotPluginOAuth: Record<string, string> = {};
-copilotPluginOAuth["api-auth"] = "oauth";
-
-const env = Object.assign({}, process.env);
-env["API_COPILOT_PLUGIN"] = "true";
-env["DEVELOP_COPILOT_PLUGIN"] = "true";
-
-const options = {
-  skipValidate: true,
-  skipErrorMessage: "No elements found in the manifest",
-};
+class CopilotPluginOAuthTestCase extends CopilotPluginCommonTest {}
 
 new CopilotPluginOAuthTestCase(
-  Capability.CopilotPluginFromScratch,
   27569691,
   "huimiao@microsoft.com",
-  ["function"],
-  ProgrammingLanguage.TS,
-  options,
-  copilotPluginOAuth,
-  env
+  "oauth",
+  ProgrammingLanguage.TS
 ).test();
